requirements.txt
asgiref==3.4.1
astroid==2.4.2
certifi==2022.12.7
charset-normalizer==3.0.1
dj-database-url==0.5.0
Django==3.2.8
django-heroku==0.3.1
djangorestframework==3.12.4
docopt==0.6.2
gunicorn==20.1.0
idna==3.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pipreqs==0.4.11
psycopg2==2.9.5
psycopg2-binary==2.9.3
pylint==2.6.0
python-dotenv==0.20.0
pytz==2021.3
requests==2.28.2
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.14
whitenoise==6.2.0
wrapt==1.12.1
yarg==0.1.9
